feat: add LogHeader component
